Project Update:
Removed the default template assets and started replacing with required assets for project. For now includes a basic home page with the other 4 pages just blank for now. .css files have been set up to be properly shared across the correct pages as well. Need to populate pages with info and stylize the pages with .css. Will add .js files later if needed.