Stylized Homepage and About Me.
Started working on the stylization for both the homepage and about me. I have the basic set up for shared aspects across the pages so I just need to fill in the other pages. Includes a new background for the homepage, and allows the user to download my resume (Some information removed for safety reasons).